backend/requirements.txt
Django==5.2.6
django-jazzmin==3.0.1
python-dotenv==1.1.0
Pillow==12.0.0
requests==2.32.5
djangorestframework
django-cors-headers
djoser
djangorestframework-simplejwt
django-filter
sslcommerz-lib
